+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Changelog
 v2.5.3
 ------
 * Added convenience methods for generating conditional logic.
+* Added a method which can check if the class being made has any unimplemented abstract methods.
 
 v2.5.2 (2023-11-27)
 ------